producer_simulator/requirements.txt
confluent_kafka==2.3.0
Django==4.2.3
djangorestframework==3.14.0
